requirements.txt
flask
flask_sqlalchemy
mysqlclient
yfinance
ta
python-dotenv
pandas
numpy
flask-cors
gunicorn
mysql-connector-python